add webmin tool script
Webmin is a web-based system administration tool for Unix-like systems. It provides an easy alternative to command line system administration and can be used to manage various aspects of a system, such as users and services, through the use of the provided Webmin modules. If you want to manage your own server but you are uncomfortable with the command line, Webmin is a good tool to help you get started.